Scene Root Groundwork
Started support for scene root system. Currently, any prop can be "parented" to any other prop, with its location and rotation defined relative to that prop. "Parents" can be reassigned, and a system is in place to prevent infinite loops. Rooted props automatically move when their "parent" is moved with setposang.