Dialogcmds Clean up
Removed commented old local/world code since it's definitely unneeded.

Technically adds start of camera bounds checking but it's currently dummied out.

Also makes it so setcamroot doesn't need a location defined, if it's meant to stay at the same offset it was previously, or to be used as a dummy command for generating offsets.