Color trick
Using the fact that the dialog options can't render special text attributes to hide action indicators in player-chosen texts (i.e. "(exit)" text will appear on the option to pick, but if picked it won't show in the following dialog)

Added to already-converted scripts, will be adding it to the rest as conversions continue.