zSnap adjustment
zSnap now assumes a Z offset provided to it is an adjustment to its hit position on the floor. This allows a prop to be adjusted to not partially clip into the floor if its origin isn't at its bottom.

Adjusted previously converted scripts to take this into account.

(also added shots to bartransition)