Cellist Event 1
And now, the reason for all of these changes today. Drumroll please.

Added the Good Shit to the scene, in addition to its conversion.

Tweaked script some more than that as well, mostly in the ending. "Safely skip" for the drug scene no longer flashes white like the default scene does, and the player is down for the count at the end.