setlocale update
Setlocale can be provided a list of potential backup locales to try. It will attempt them in order from first listed to last, with the first one successfully found being used. If none are found, it still roots to the map origin like before.
Current limit is 8 locales.

Updated converted scripts with this functionality, mostly doing backups for "_solo" and non-"_solo" locales.