Layers System
Added on top of the zSnap/Gravity System is the layers system. Scenes are able to utilize layers where things are to be set, rather that needing to depend solely on zSnap to take care of them. This is useful for concavities in the scene locale that need to be accounted for (i.e., being able to have props both on the stools at the bar, and props sitting on top of the bar awning).

Layers are all relative to the locale entity. All scenes have the base layer 1, which is set at 0 (i.e. the Z of the locale entity)

An entity/entities can be set to another layer with the `setlayer [layer number] [entity1 entity2 ... entityn]` command