show updates
Updated rest of scripts to use new show functionality. For consistency's sake, but more than just in the script itself. This helps make sure there's no slight visual hitches from the speaker being set at a different time than the text box is set to show. Some were before it, some were after it prior to the update.

Will be testing a reversion to the quick "nil" fix to see if this helps solve it (in scenes, anyway. Still likely to get a nil box on occasion when talking to a cat)