setlook command
(Now that it's inexplicably working for some reason, tried this like a month ago with the same exact damn code and it didn't work) Added setlook command. This adjusts a prop's head by the given angle. Format is `*setlook name [pitch] [yaw] [roll]*`. Any portion of the angle not set is assumed zero (i.e. reset).

Going to go through and remove all the MJ leans that the player does in order to look at a cat. That said:

Added to several scenes. Also includes some other consistency fixes that slipped through the cracks previously.