Further edict limit protection
Lots of IsValid() checks/some extra logic to make sure we're not throwing a lua error/breaking shit and fucking up getting out
Functionality changes to make strictly necessary entities more likely to get spawned over superficial ones in near edict limit situations.

If a bus is unable to be summoned due to edict limit, bus summoners will fall back to a "voting" system. Namely:
Any player attempting to use a summoner will count as a vote to leave.
If a majority of player-owned summoners are currently voting to leave, the map is left.

If not a single player is able to get a bus summoner... then we're still stuck on the map for now. Sorry.