setcamtarget
Adds a new camera control option to scripts, and subsequent command to set/clear it.
`*setcamtarget [targetname] [offset vector]*`
causes the camera's view to follow targetname's position, with a vector for an offset from it (so we aren't always staring at characters' feet with it)
Both parameters are optional, if targetname is left off we update the offset to the current target. If vector is left off we set the new target with the old offset. If both are left off (or `setcamtarget nil` is used) the target is cleared and camera returns to normal functionality.
Note that the camera's view won't be reset by itself when the target is cleared. Though if the camera is in the process of tweening (either itself or its root) it *will* snap back to where it would be in that tween.