BLOCKSTART/END validation
Adds validation checks to exec block formatting. Can identify if the blockstart was missing a blockend, or if plaintext broke the formatting of the block, and reports them.

Also does a little more cleanup on processed strings to try to help reduce the size that we're needing to send/store.