Local Macros
Adds the ability to define per-script macros.
These are added by putting
`MACRO [macro name](optional parameters) [macro commands]`
before any branches in the script. They function fully the same as global macros, and are always applied first, so they can use global macros in them as well.
Currently untested, but, they should *in theory* also be able to override a global macro for the script, but this probably shouldn't be relied on even if it's true.