Handle ./models/... props
So apparently it's possible for some prop models to be listed as "./models/..." in their filepath. This works fine for collecting them and spawning them at the prop chute, but breaks the void shell being made for them.

To help offset the cost of the new string processing to be able to handle these, also using it to remove the "mdl" from the end of the strings, rather than leave that to the individual functions.