Barhopping
Added the basis for being able to switch to a different hub map from within said map, from now on referred to as Barhopping!
Hub maps can use values set in the selector to define their ideal setup for relevant server convars (namely, jazz_hub which is automatically set to the map's name, and jazz_trolley which can define a non-standard trolley appearance script).
When `jazz_barhop_allow` is enabled and these "ideal" settings don't match the server's settings, a new dialog option will appear for the Bartender. Super Admins can use this to easily match the server's settings to the map's.