jazz_vote_percentage
Adds a new convar, jazz_vote_percentage. Its value is used by the voting podiums for selecting endings. The players who voted to start an ending must have made at least jazz_vote_percentage% of the server's money this reset for the ending to begin.
Mainly for public servers, this prevents a lone, new player from joining a server and making an ending happen. Default value is 0.51 (51% of total money). Keep in mind that all players on the server must still vote for the ending to be selected.