jazz_logic_purchase
Added new entity which allows mappers to make arbitrary Jazzy money purchases available for players, and has outputs for an attempted purchase, and then successful, unsuccessful, or (optionally) partially-successful purchases.
Note that the player in question must be the activator for the logic chain, and the map in question has to be set as the server's hub for any purchases to be able to be made. In addition:

Also adds new clientconvar, `jazz_money_protection`. Sets the player's comfort level with the new entity. Set to 0 to allow all purchases. Set to 1 (default) to only allow purchases made through entities that (in theory, probably have to test more) require the player to +USE them to activate the purchase. Set to 2 to disable all purchases (that aren't the bartender's shop).

The entity will warn players when an attempt to make a purchase that isn't allowed is blocked. (Yes I am aware that if someone really wanted to, they could fuck us up regardless of any protections made here. Still should at least try to block the low hanging fruit and hope we haven't pissed anyone off enough for them to try harder.)