Properly use backup addon list
Previously, if the provided host returned something that wasn't a properly formatted addon list, or if it was a bad request but the cache was *also* malformed because the previous host was bad, the game would fail to actually use the backup list in the files. It now properly checks that the addon list has *something* valid in it and uses the backup if it doesn't.

Also, throws some non-halting errors so it's easier for people to tell when it's borked.